chore: Move nextSaturday
